style(whatislove-dev): rename packages to modules wd-291
