+++ packages/whatislove-dev/src/scripts/globals.d.ts
@@ -1,4 +1,4 @@
-import { WhatisloveMath } from '~/libs/packages/whatislove-math/whatislove-math.js'
+import { WhatisloveMath } from '~/libs/modules/whatislove-math/whatislove-math.js'
 
 declare global {
 	var WhatisloveMath: WhatisloveMath
