+++ packages/whatislove-dev/src/scripts/index.js
@@ -1,6 +1,6 @@
-import { storage } from '~/libs/packages/storage/storage.js'
-import { WhatisloveMath } from '~/libs/packages/whatislove-math/whatislove-math.package.js'
-import { timelineApi } from '~/packages/timeline/timeline.js'
+import { storage } from '~/libs/modules/storage/storage.js'
+import { WhatisloveMath } from '~/libs/modules/whatislove-math/whatislove-math.package.js'
+import { timelineApi } from '~/modules/timeline/timeline.js'
 import { Main } from '~/pages/main/main.js'
 
 let main = new Main({
