+++ packages/whatislove-dev/src/scripts/libs/modules/http/libs/enums/http-method.enum.js
