+++ packages/whatislove-dev/src/scripts/libs/modules/http/libs/exceptions/http-error/http-error.exception.js
