+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline-form/helpers/get-transformed-timeline/get-transformed-timeline.helper.js
@@ -1,4 +1,4 @@
-import { TimelineCreatePayload } from '~/packages/timeline/timeline.js'
+import { TimelineCreatePayload } from '~/modules/timeline/timeline.js'
 
 /**
  * @param {TimelineCreatePayload} timeline
