+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline/libs/components/timeline-form/timeline-form.js
@@ -1,5 +1,5 @@
 import { getFormValues, initDebounce } from '~/libs/helpers/helpers.js'
-import { TimelineFilter } from '~/packages/timeline/timeline.js'
+import { TimelineFilter } from '~/modules/timeline/timeline.js'
 
 import { DEBOUNCE_DELAY } from './libs/constants/constants.js'
