+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline/libs/components/timeline-list/libs/helpers/get-timeline-template/get-timeline-template.helper.js
@@ -1,5 +1,5 @@
 import { getFormattedDate, getStringWitCheck } from '~/libs/helpers/helpers.js'
-import { Timeline } from '~/packages/timeline/timeline.js'
+import { Timeline } from '~/modules/timeline/timeline.js'
 
 /**
  * @param {Timeline} timeline
