+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline/libs/components/timeline-list/libs/helpers/get-timeline-templates/get-timeline-templates.helper.js
@@ -1,4 +1,4 @@
-import { Timeline } from '~/packages/timeline/timeline.js'
+import { Timeline } from '~/modules/timeline/timeline.js'
 
 import { getTimelineTemplate } from '../get-timeline-template/get-timeline-template.helper.js'
