+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline/libs/components/timeline-list/timeline-list.js
@@ -1,4 +1,4 @@
-import { Timeline } from '~/packages/timeline/timeline.js'
+import { Timeline } from '~/modules/timeline/timeline.js'
 
 import { getTimelineTemplates } from './libs/helpers/helpers.js'
