+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline/libs/helpers/check-is-timeline-suit/check-is-timeline-suit.helper.js
@@ -1,4 +1,4 @@
-import { Timeline, TimelineFilter } from '~/packages/timeline/timeline.js'
+import { Timeline, TimelineFilter } from '~/modules/timeline/timeline.js'
 
 /**
  * @param {Timeline} timelineItem
