+++ packages/whatislove-dev/src/scripts/pages/main/libs/components/timeline/libs/helpers/get-suit-timelines/get-suit-timelines.helper.js
@@ -1,4 +1,4 @@
-import { Timeline, TimelineFilter } from '~/packages/timeline/timeline.js'
+import { Timeline, TimelineFilter } from '~/modules/timeline/timeline.js'
 
 import { checkIsTimelineSuit } from '../check-is-timeline-suit/check-is-timeline-suit.helper.js'
